# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cabales API
  version: 1.2.0
  license: { name: ISC, identifier: ISC }
  description: >-
    API de autenticación, grupos, eventos, gastos exactos, liquidaciones, fondos, presupuestos, documentos, OCR, Cabudas, estadísticas, avisos, logros y privacidad. Importes en centavos.
    Todas las respuestas /api/v1 usan Cache-Control private, no-store; health y ready usan no-store.
servers:
  - url: /
...
paths:
...
  /api/v1/ocr/jobs:
    get:
      tags: [OCR]
      summary: 'Trabajos OCR propios'
      operationId: listOcrJobs
      parameters:
        - name: documentId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/OcrJob' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [OCR]
      summary: 'Encola la lectura de un comprobante'
      operationId: createOcrJob
      description: 'Requiere acceso de lectura al documento (PDF o imagen). 503 OCR_PROVIDER_DISABLED si no hay proveedor configurado. OCR_PROVIDER=local solo existe en desarrollo/pruebas y la respuesta marca provider=local. El OCR solo propone datos; nunca modifica gastos.'
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [documentId]
              properties:
                documentId:
                  type: string
                  format: uuid
      responses:
        '202':
          description: 'Aceptado; el procesamiento es asíncrono'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/OcrJob' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/ocr/jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [OCR]
      summary: 'Estado del trabajo; proposal solo cuando SUCCEEDED'
      operationId: getOcrJob
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/OcrJob' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/ocr/jobs/{jobId}/retry:
    parameters:
      - $ref: '#/components/parameters/JobId'
    post:
      tags: [OCR]
      summary: 'Reintenta un trabajo fallido dentro del máximo de intentos'
      operationId: retryOcrJob
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '202':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/OcrJob' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/ocr/jobs/{jobId}/confirm:
    parameters:
      - $ref: '#/components/parameters/JobId'
    post:
      tags: [OCR]
      summary: 'Confirma la propuesta vinculando el documento a un gasto revisado'
      operationId: confirmOcrJob
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expenseId]
              properties:
                expenseId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/OcrJob' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /api/v1/privacy/requests:
    get:
      tags: [Privacy]
      summary: 'Solicitudes ARCO-POL propias'
      operationId: listPrivacyRequests
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PrivacyRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Privacy]
      summary: 'Crea una solicitud de acceso, rectificación, supresión, oposición o portabilidad'
      operationId: createPrivacyRequest
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PrivacyRequestInput' }
      responses:
        '201':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/PrivacyRequest' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/privacy/requests/{requestId}:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      tags: [Privacy]
      summary: 'Detalle de una solicitud propia'
      operationId: getPrivacyRequest
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/PrivacyRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/privacy/requests/{requestId}/confirm:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      tags: [Privacy]
      summary: 'Confirma la solicitud; la supresión exige contraseña'
      operationId: confirmPrivacyRequest
      description: 'ACCESS y PORTABILITY quedan COMPLETED con exportación temporal. ERASURE anonimiza la cuenta, transfiere la propiedad de grupos compartidos, elimina credenciales, sesiones y documentos propios y limpia cookies. RECTIFICATION y OBJECTION pasan a IN_PROGRESS para atención manual.'
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: []
              properties:
                password:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/PrivacyRequest' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/privacy/requests/{requestId}/cancel:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      tags: [Privacy]
      summary: 'Cancela una solicitud pendiente'
      operationId: cancelPrivacyRequest
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/PrivacyRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/privacy/requests/{requestId}/export:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      tags: [Privacy]
      summary: 'Descarga la exportación JSON de datos propios'
      operationId: exportPrivacyRequest
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    description: 'Perfil, membresías, eventos, gastos con su parte, transferencias, fondos, documentos (metadatos), avisos, solicitudes y actividad propia. Sin secretos ni datos personales de terceros.'
          headers:
            Content-Disposition:
              schema:
                type: string
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cabales_session
      description: Nombre predeterminado configurable mediante COOKIE_NAME.
    csrfCookie:
      type: apiKey
      in: cookie
      name: cabales_session_csrf
      description: Cookie de doble envío cuyo prefijo sigue COOKIE_NAME; /auth/me valida su hash antes de devolverla.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Debe coincidir con csrfCookie y con el hash ligado a la sesión.
  parameters:
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    BadRequest:
      {
        description: Entrada inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Unauthorized:
      {
        description: Sesión ausente o inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Forbidden:
      {
        description: CSRF o permiso inválido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    NotFound:
      {
        description: Recurso no visible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Conflict:
      {
        description: Estado concurrente o idempotencia incompatible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
    UnprocessableEntity:
      {
        description: Invariante de negocio incumplida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    TooManyRequests:
      {
        description: Límite de solicitudes excedido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
    ServiceUnavailable:
      {
        description: PostgreSQL no disponible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
...
    OcrProposal:
      type: object
      required: [items]
      properties:
        merchant:
          type: [string, 'null']
        totalCents:
          type: [integer, 'null']
        currency:
          type: [string, 'null']
        occurredAt:
          type: [string, 'null']
          format: date-time
        items:
          type: array
          items:
            type: object
            required: [description, amountCents]
            properties:
              description:
                type: string
              amountCents:
                type: integer
        confidence:
          type: number
          minimum: 0
          maximum: 1
    OcrJob:
      type: object
      required:
        [
          id,
          documentId,
          status,
          attempts,
          errorCode,
          createdAt,
          startedAt,
          finishedAt,
          confirmedAt,
          confirmedExpenseId,
          proposal,
          maxAttempts,
          canRetry,
          provider,
        ]
      properties:
        id:
          type: string
          format: uuid
        documentId:
          type: [string, 'null']
          format: uuid
        status:
          enum: [PENDING, PROCESSING, SUCCEEDED, FAILED]
        attempts:
          type: integer
        errorCode:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        finishedAt:
          type: [string, 'null']
          format: date-time
        confirmedAt:
          type: [string, 'null']
          format: date-time
        confirmedExpenseId:
          type: [string, 'null']
          format: uuid
        proposal:
          oneOf:
            - $ref: '#/components/schemas/OcrProposal'
            - type: 'null'
        maxAttempts:
          type: integer
        canRetry:
          type: boolean
        provider:
          type: string
...
    PrivacyRequest:
      type: object
      required:
        [
          id,
          type,
          status,
          reason,
          createdAt,
          updatedAt,
          confirmedAt,
          cancelledAt,
          completedAt,
          exportExpiresAt,
          resultSummary,
          exportAvailable,
          requiresPassword,
        ]
      properties:
        id:
          type: string
          format: uuid
        type:
          enum: [ACCESS, RECTIFICATION, ERASURE, OBJECTION, PORTABILITY]
        status:
          enum: [PENDING, IN_PROGRESS, COMPLETED, REJECTED, CANCELLED]
        reason:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        confirmedAt:
          type: [string, 'null']
          format: date-time
        cancelledAt:
          type: [string, 'null']
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        exportExpiresAt:
          type: [string, 'null']
          format: date-time
        resultSummary:
          type: [object, 'null']
        exportAvailable:
          type: boolean
        requiresPassword:
          type: boolean
    PrivacyRequestInput:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          enum: [ACCESS, RECTIFICATION, ERASURE, OBJECTION, PORTABILITY]
        reason:
          type: string
          maxLength: 1000